Use Unicode stars rather than graphical stars for ratings.
I couldn’t find stars.svg being used anywhere else so I removed the
file.

This has the original fix from PR #1482, and fixes the issue with star
selection in the write-a-review modal that was found in #1483.

However, I’m unable to test the issue that was mentioned in  #1483
regarding the star sizes being too small on FF on Windows, so this PR
is missing that fix.